+++ packages/core/src/protocols/schemas/project.ts
@@ -73,6 +73,11 @@ export interface ProjectSchema {
   uniConfig?: UniConfig;
 
   /**
+   * （Web、H5）应用全局配置
+   */
+  globals?: GlobalConfig;
+
+  /**
    * 标记
    */
   __VTJ_PROJECT__?: boolean;
@@ -94,7 +99,7 @@ export interface ProjectSchema {
 }
 
 /**
- * 项目配置信息
+ * 项目母版配置信息
  */
 export interface ProjectConfig {
   /**
@@ -178,4 +183,54 @@ export interface UniConfig {
    * onExit	监听应用退出
    */
   onExit?: JSFunction;
+}
+
+/**
+ * 应用全局配置项
+ */
+export interface GlobalConfig {
+  /**
+   * 应用全局样式
+   */
+  css?: string;
+
+  /**
+   * 应用全局状态
+   */
+  store?: JSFunction;
+
+  /**
+   * 权限控制插件配置项
+   */
+  access?: JSFunction;
+
+  /**
+   * 应用增强函数
+   */
+  enhance?: JSFunction;
+
+  /**
+   * 请求工具配置项
+   */
+  axios?: JSFunction;
+
+  /**
+   * 请求拦截器
+   */
+  request?: JSFunction;
+
+  /**
+   * 响应拦截器
+   */
+  response?: JSFunction;
+
+  /**
+   * 前置路由守卫
+   */
+  beforeEach?: JSFunction;
+
+  /**
+   * 后置路由守卫
+   */
+  afterEach?: JSFunction;
 }