+++ packages/designer/src/components/hooks/index.ts
@@ -17,3 +17,4 @@ export * from './useOpenApi';
 export * from './useTemplates';
 export * from './useCheckVersion';
 export * from './useAI';
+export * from './useWorkspace';