+++ packages/designer/src/components/hooks/useDesigner.ts
@@ -4,7 +4,8 @@ import {
   type ApiSchema,
   type MetaSchema,
   type ProjectConfig,
-  type UniConfig
+  type UniConfig,
+  type GlobalConfig
 } from '@vtj/core';
 import { useEngine, type DesignHelper } from '../../framework';
 
@@ -14,11 +15,20 @@ export function useDesigner(
   apis: Ref<ApiSchema[]>,
   meta: Ref<MetaSchema[]>,
   config: Ref<ProjectConfig>,
-  uniConfig: Ref<UniConfig>
+  uniConfig: Ref<UniConfig>,
+  globals: Ref<GlobalConfig>
 ) {
   const engine = useEngine();
 
-  engine.simulator.init(iframe, dependencies, apis, meta, config, uniConfig);
+  engine.simulator.init(
+    iframe,
+    dependencies,
+    apis,
+    meta,
+    config,
+    uniConfig,
+    globals
+  );
 
   const designer = computed(() => engine.simulator.designer.value);