+++ packages/designer/src/components/hooks/useProject.ts
@@ -3,7 +3,6 @@ import { useEngine } from '../../framework';
 export function useProject() {
   const engine = useEngine();
   const project = computed(() => engine.project.value);
-
   return {
     engine,
     project