+++ packages/uni/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.13.0](https://gitee.com/newgateway/vtj/compare/@vtj/uni@0.12.70...@vtj/uni@0.13.0) (2025-08-06)
+
+**Note:** Version bump only for package @vtj/uni
+
+
+
+
+
 ## [0.12.70](https://gitee.com/newgateway/vtj/compare/@vtj/uni@0.12.69...@vtj/uni@0.12.70) (2025-08-01)
 
 **Note:** Version bump only for package @vtj/uni